requirements.txt
pandas>=2.0.0
google-cloud-bigquery>=3.11.0
google-cloud-bigquery-storage>=2.22.0
db-dtypes>=1.1.1
requests>=2.31.0
streamlit>=1.28.0
plotly>=5.17.0
python-dotenv>=1.0.0
mysql-connector-python>=8.1.0
pyarrow>=13.0.0
